pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flp"
version = "0.1.0"
description = "Federated Learning Playground — a simulation-first experimentation framework for distributed AI training"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "meetnishant", email = "meetnishant@users.noreply.github.com" }]
keywords = ["federated-learning", "machine-learning", "simulation", "distributed-ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "matplotlib>=3.8.0",
    "click>=8.1.0",
    "numpy>=1.26.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.scripts]
flp = "flp.cli:main"

[project.urls]
Repository = "https://github.com/meetnishant/FederatedLearningPlayground"

[tool.hatch.build.targets.wheel]
packages = ["src/flp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
